requirements.txt
alabaster==0.7.8
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
#Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
#chrome-gnome-shell==0.0.0
cryptography==3.3.2
#cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
distro-info==1.0
docutils==0.16
duplicity==0.8.17
#fasteners==0.14.1
#feedparser==5.2.1
#fuse-python==1.0.2
#future==0.18.2
#gpg===1.14.0-unknown
#gyp==0.1
#html5lib==1.1
#httplib2==0.18.1
#idna==2.10
#imagesize==1.2.0
#invoke==1.4.1
#Jinja2==2.11.3
#lockfile==0.12.2
#louis==3.16.0
#lxml==4.6.3
#Mako==1.1.3
#MarkupSafe==1.1.1
#meson==0.56.2
#monotonic==1.5
#musicbrainzngs==0.7.1
#mutagen==1.45.1
#numpy==1.23.4
#oauthlib==3.1.0
#olefile==0.46
#packaging==20.9
pandas==1.5.1
#paramiko==2.7.2
#pexpect==4.8.0
#Pillow==8.1.2
#psutil==5.8.0
#pycairo==1.16.2
#pycups==2.0.1
#pycurl==7.43.0.6
#Pygments==2.7.1
#PyGObject==3.38.0
#pyinotify==0.9.6
PyJWT==1.7.1
#pylibacl==0.6.0
#PyNaCl==1.4.0
pyodbc==4.0.35
#pyparsing==2.4.7
#PyQt5==5.15.2
#PyQt5-sip==12.8.1
#PySimpleSOAP==1.16.2
#pysmbc==1.0.23
#python-apt==2.2.1
#python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-decouple==3.6
python-dotenv==0.21.0
python-pam==1.8.4
python-tds==1.11.0
#python-xapp==2.0.2
python-xlib==0.29
pytz==2021.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
#quodlibet==4.3.0
#reportbug==7.10.3+deb11u1
requests==2.25.1
roman==2.0.0
scour==0.38.2
setproctitle==1.2.1
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.4.3
tinycss==0.4
tinycss2==1.0.2
tornado==6.1
unattended-upgrades==0.1
urllib3==1.26.5
vboxapi==1.0
webencodings==0.5.1
xdg==5
